Refactored refreshable label
