--- src/components/UI/refreshable_label.js
@@ -1,92 +1,0 @@
-import { keyframes } from '@emotion/core';
-import _ from 'underscore';
-import PropTypes from 'prop-types';
-import React from 'react';
-import ReactTimeout from 'react-timeout';
-import styled from '@emotion/styled';
-
-const yellowfade = keyframes`
-  from {
-    background: #e8d986;
-  }
-  to {
-    background: transparent;
-  }
-`;
-
-const Wrapper = styled.div`
-  display: inline-block;
-  line-height: 1.7;
-  border-radius: 2px;
-  margin-left: -5px;
-  padding-left: 5px;
-  padding-right: 5px;
-  &.changed {
-    animation: ${yellowfade} 2s ease;
-  }
-`;
-
-/**
- * RefreshableLabel is an inline-block HTML container
- * that is used to temporarily highlight its content.
- *
- * It is typically used to indicated that its content
- * has changed. To detect a change, it provides a property
- *
- *   dataItems (Array)
- *
- * which is an array of arbitrary values. When this
- * array changes, visual highlighting is triggered.
- *
- * Note: props.setTimeout is added via ReactTimeout.
- */
-class RefreshableLabel extends React.Component {
-  state = {
-    dataItems: this.props.dataItems,
-    changed: false,
-  };
-
-  static getDerivedStateFromProps = (nextProps, prevState) => {
-    var d = _.difference(prevState.dataItems, nextProps.dataItems);
-    if (_.size(d) === 0) {
-      return null;
-    }
-
-    return {
-      dataItems: nextProps.dataItems,
-    };
-  };
-
-  componentDidUpdate(prevProps, prevState) {
-    if (
-      prevState.dataItems !== null &&
-      prevState.dataItems !== this.state.dataItems
-    ) {
-      if (prevState.dataItems !== null) {
-        this.setState({
-          changed: true,
-        });
-
-        this.props.setTimeout(() => {
-          this.setState({ changed: false });
-        }, 5000);
-      }
-    }
-  }
-
-  render() {
-    return (
-      <Wrapper className={this.state.changed ? 'changed' : null}>
-        {this.props.children}
-      </Wrapper>
-    );
-  }
-}
-
-export default ReactTimeout(RefreshableLabel);
-
-RefreshableLabel.propTypes = {
-  children: PropTypes.object,
-  dataItems: PropTypes.array,
-  setTimeout: PropTypes.func,
-};
